+++ pyproject.toml
@@ -43,6 +43,7 @@ tensorflow-probability = "^0.23.0"
 jaxtyping = "^0.2.24"
 beartype = "^0.18.5"
 deprecated = "^1.2.14"
+penzai = "^0.2.2"
 treescope = "^0.1.4"
 # Numpy <2.0.0 is due to tfp issues: https://github.com/tensorflow/probability/issues/1814
 # remove this constraint when that issue is closed.
